Update canister-sdk deps

+++ ic-stable-structures/src/storage/structures/log.rs
@@ -32,12 +32,12 @@ impl<T: Storable> StableLog<T> {
     }
 
     /// Returns reference to value stored in stable memory.
-    pub fn get(&self, index: usize) -> Option<T> {
+    pub fn get(&self, index: u64) -> Option<T> {
         self.get_inner()?.get(index)
     }
 
     /// Updates value in stable memory.
-    pub fn append(&mut self, value: T) -> Result<usize> {
+    pub fn append(&mut self, value: T) -> Result<u64> {
         let canister_id = ic::id();
         let index = match self.data.entry(canister_id) {
             Entry::Occupied(mut entry) => entry
@@ -57,7 +57,7 @@ impl<T: Storable> StableLog<T> {
     }
 
     /// Number of values in the log.
-    pub fn len(&self) -> usize {
+    pub fn len(&self) -> u64 {
         self.get_inner()
             .map(|inner| inner.len())
             .unwrap_or_default()
